requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==10.0.0
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai-whisper==20230314
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
sympy==1.12
tiktoken==0.3.1
tk==0.1.0
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
